refactor(core): relocate session_type() from _type_enums to _type_helpers
Moves the runtime env-reader function session_type() out of _type_enums.py
(a pure StrEnum definitions file) into _type_helpers.py where runtime helpers
belong. Eliminates private constant duplicates (_SESSION_TYPE_ENV_VAR,
_HEADLESS_ENV_VAR) in favour of the public constants from _type_constants.py.
Adds a module-placement test that asserts the relocation.
