+++ tests/core/test_session_type.py
@@ -139,3 +139,23 @@ def test_session_type_fleet_constant_matches_enum():
     from autoskillit.core._type_constants import SESSION_TYPE_FLEET
 
     assert SessionType.FLEET.value == SESSION_TYPE_FLEET
+
+
+# ---------------------------------------------------------------------------
+# Group D — Module placement
+# ---------------------------------------------------------------------------
+
+
+def test_session_type_is_defined_in_type_helpers():
+    """session_type() must live in _type_helpers, not _type_enums."""
+    import importlib
+    import inspect
+
+    from autoskillit.core._type_helpers import session_type as fn_helpers
+
+    type_enums = importlib.import_module("autoskillit.core._type_enums")
+    assert not hasattr(type_enums, "session_type"), (
+        "session_type must not be defined in _type_enums after relocation"
+    )
+
+    assert inspect.getmodule(fn_helpers).__name__ == "autoskillit.core._type_helpers"
